[FEATURE] New element "Password with confirmation"
This adds a new FormElement "TYPO3.Form:PasswordWithConfirmation"
that displays two password fields.

Technical details:
The element checks whether both fields are equal and add a validation error
to the corresponding processing rule if that's not the case.
Otherwise the element value is changed to the password so that additional
regular validators can be used to enforce length and complexity of the
password.

Change-Id: I80c68bdd04b89eaa9c4687798291246a0e88123f