[FEATURE] Support identifiers with special characters
Currently it's not possible to add FormElements with non alpha
numeric characters in the identifier.
This is required in order to specify property paths (foo.bar.baz)
and to add elements that should map to internal arguments (__foo).

This change removes the strict regex check from the FormElement
constructor and instead makes sure that
FormElement::getUniqueIdentifier() returns an identifier without
special characters so it can be used as id attributes in forms for
example.

Change-Id: I99f8a9bf3ef13147e1baef655a6900a0761d085c
Resolves: #36722