[!!!][BUGFIX] Default values can't be set recursively
Currently, if you use dots in the form element identifier in
order to enable property mapping over multiple fields default
values are no displayed as expected.
This change fixes this by storing all default values in the
FormDefinition as multidimensional array.

This is a breaking change only if you set a default value to
an form element before attaching the element to a form.
If you use the recommended syntax createElement() that will
work as before.

Change-Id: I0754cbbe70340b7f443a0ad368b292863bd065dc
Resolves: #36819