[!!!][FEATURE] Make class attributes of elements configurable
Currently class attributes are hard coded in the provided Form
Element templates.
This change extracts those classes to the default preset, so
you can change them by adjusting the preset or by creating a
custom preset.

This is not really a breaking change, most classes stay the
same and you won't have to adjust configuration or templates.
There are two exceptions:
1. We don't render an additional div container around the
single checkbox as that's deprecated with Twitter Bootstrap 2.x
If you need that, you have to override the Checkbox template.
2. Setting the class attribute of the StaticText paragraph
now works in sync with all the other elements. If you used the
property "class" to set that, you can replace that property by
"elementClassAttribute" and everything works as before.

Resolves: #37728

Change-Id: I2795d45e0b359906223cc73cb9011bbbdc59b5fc