[FEATURE] Provide a better Exception for PropertyMapping
If the PropertyMapper fails to map some properties it
can be hard to understand what happened, because
the PropertyMapper doesn't get the whole PropertyPath
from the ProcessingRules.
To better understand what happens this change catches
the Property\Exception and throws a new Exception
with the propertyPath that was tried to be mapped.

Example:
http://dl.dropbox.com/u/314491/Screenshots/09.png

Without this Change only the 2 nested Exceptions
would have been thrown.

Change-Id: I6a96599ee6cdab25a3cab379d27dbaaed96b6f12