[FEATURE] Skip unknown FormElements
With this change FormElements without definition in the current
preset can be skipped while rendering.

In ``previewMode`` (e.g. inside the FormBuilder) instead of an
exception an error message is now rendered in the form::

 <div class="typo3-form-unknown-element" data-element="element/path">
   <em>Unknown Form Element "<type>"</em>
  </div>

During rendering of the Form unknown Form Elements still throw
an exception by default for backwards compatibility reasons.
But this behavior can be changed for single presets with the
new rendering option ``skipUnknownElements``::

TYPO3:
  Form:
    presets:
      'somePresetName':
        formElementTypes:
          'TYPO3.Form:Base':
            renderingOptions:
              skipUnknownElements: TRUE

With this option set, unknown Form Elements won't be rendered
at all.

Change-Id: I1746d636d013c8a345f8dc668037c64c656894ed
Resolves: #34198