[TASK] Support boolean superType-definitions for FormElements
It is misleading and error-prone that the FormElement superTypes-definition
works different than the ContentRepository superTypes.

This patch aligns the FormElement superTypes with the syntax of ContentRepository
superTypes. The old syntax is supported aswell if the given value is
not of type Boolean.

previous format::

  'Some.Package:SomeType':
    superTypes: ['Some.OtherPackage:SomeSuperType']

new format::

  'Some.Package:SomeType':
    superTypes:
      'Some.OtherPackage:SomeSuperType': TRUE

Change-Id: I8abd61e3ddbe1266a256b5cac988c3d14491b2bc
Releases: master, 2.0
Related: NEOS-1304