[TASK] Scroll form into view upon submission
Currently the action URI of a form is always set to the requested URI.
However when a form is at the bottom of a page and gets submitted, the
visitor will end up at the top of the page and has to scroll down to
see if everything is correct.

This patch fixes this behavior by making the current ``formIdentifier``
the section (aka anchor, fragment identifier) of the form action URI.

The custom ``FormViewHelper`` didn't respect the ``section``
argument. This bug is fixed with this patch as well.

If you want to disable this behavior, you can override the ``Form.html``
template.

Change-Id: I76d208f9bd28dd04382ee2f681c6768ad1856a2a
Related: NEOS-1353
Releases: master, 2.0