FEATURE: Configurable Form Element render callbacks
Adds an API method `registerRenderCallback()` to the `FormRuntime`
that allows for influencing the rendering of rendered elements.

Usage:

    $formRuntime->registerRenderCallback(function (string $output, RootRenderableInterface $renderable) {
        return $renderable->getIdentifier() . ': ' . $output;
    });

This simple example will render the Element identifier in front of every form
element, independantly from the configured Form (Element) Renderer.

Background:
The rendering of Form Elements can already be configured and altered
thoroughly, but this allows for 3rd party integration. For example we can
use this to wrap each Form Element using the `ContentElementWrappingService`
in order to render the required wraps in the Neos Backend.