FEATURE: EmailFinisher file attachments
Extends the `EmailFinisher` to be able to send files
(uploaded via the form or static ones) as attachment(s).

Usage:

```yaml
finishers:
  -
    identifier: 'Neos.Form:Email'
    options:
      # ...
      attachAllPersistentResources: true
```

This would attach files from all FormElements with dataType `PersistentResource`.

```yaml
finishers:
  -
    identifier: 'Neos.Form:Email'
    options:
      # ...
      attachments:
        - resource: 'resource://Some.Vendor:Package/Some/Path/To/A/File'
        - formElement: 'some-upload-field'
```
This would attach a static file, referenced via the `resource` option and an uploaded
file (if the element `some-upload-field` exists in that form and contains a `PersistentResource`).

This change also improves the exception message that is thrown if the `Neos.SwiftMailer`
package is not installed.