BUGFIX: Don't proxy FormState
The `FormState` contains all the values that were
submitted to the form. It is serialized and sent via
hidden field.
This bugfix disables object management for the `FormState`
so that it can be serialized correctly.

Background:

Without this change it can happen that the FormState is
not properly deserialized and thus contains the properties
`Flow_Object_PropertiesToSerialize` & `Flow_Persistence_RelatedEntities`
that are introduced via the Flow `ProxyClassBuilder`.