Render id={form.identifier} around ConfirmationFinisher
The `<form>` is rendered always with`id={form.identifier}`
After submitting the Form `#{form.identifier}` will be added to the URL. 
If the Form is somewhere on the Page, the anchor `#{form.identifier}` will jump to the Form. 

The `Neos.Form.Builder:ConfirmationFinisher` does NOT render `id={form.identifier}` around. 
Because of the missing id, the Browser doesn't jump to the Confirmation Message and it maybe get lost somewhere on the page. 

What about wrapping an `id={form.identifier}` around the Builder:ConfirmationFinisher?